Ran into several unforseen issues that required our routes and controlers to be renamed.
What I've been refering to as traits are actually ratings so I'm going to have to rename the tests and controller. And in order to display the information correctly I need ratings to be a nested route inside users. But when I tried that it broke all path references. So for the next couple commits I'll be fixing things.